// src/app/models/thong-ke-thuy-dien.ts
export interface ThongKeThuyDien {
    thoigianxa: Date;
    ngay: string;
    gio: string;
    htl1: number;
    qvao1: number;
    luuluongnhamay1: number;
    qxaquacua1: number;
    htl2: number;
    qvao2: number;
    luuluongnhamay2: number;
    qxaquacua2: number;
    htl3: number;
    qvao3: number;
    luuluongnhamay3: number;
    qxaquacua3: number;
    htl4: number;
    qvao4: number;
    luuluongnhamay4: number;
    qxaquacua4: number;
    qvevugia: number;
    qvethubon: number;
}